+++ src/theme/index.ts
@@ -4,3 +4,4 @@ export * from './colorUtils';
 export * from './ThemeManager';
 export * from './ThemeProvider';
 export * from './useTheme';
+export * from './useThemedStyles';